requirements.txt
celery==5.3.4
Django==4.2.6
django-redis==5.4.0
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0

